+++ src/main/java/ru/m2mcom/pondnotes/service/dto/FilterPumpCleaningDTO.java
@@ -21,9 +21,9 @@ public class FilterPumpCleaningDTO implements Serializable {
     @NotNull
     private Double tempVal;
 
-    private Integer timestamp;
+    private Long timestamp;
 
-    private Integer userId;
+    private Long userId;
 
     public Long getId() {
         return id;
@@ -53,18 +53,18 @@ public class FilterPumpCleaningDTO implements Serializable {
     public void setTempVal(Double tempVal) {
         this.tempVal = tempVal;
     }
-    public Integer getTimestamp() {
+    public Long getTimestamp() {
         return timestamp;
     }
 
-    public void setTimestamp(Integer timestamp) {
+    public void setTimestamp(Long timestamp) {
         this.timestamp = timestamp;
     }
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
