+++ src/main/java/ru/m2mcom/pondnotes/service/dto/MeterReadingDTO.java
@@ -24,9 +24,9 @@ public class MeterReadingDTO implements Serializable {
     @NotNull
     private Double tempVal;
 
-    private Integer timestamp;
+    private Long timestamp;
 
-    private Integer userId;
+    private Long userId;
 
     public Long getId() {
         return id;
@@ -63,18 +63,18 @@ public class MeterReadingDTO implements Serializable {
     public void setTempVal(Double tempVal) {
         this.tempVal = tempVal;
     }
-    public Integer getTimestamp() {
+    public Long getTimestamp() {
         return timestamp;
     }
 
-    public void setTimestamp(Integer timestamp) {
+    public void setTimestamp(Long timestamp) {
         this.timestamp = timestamp;
     }
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
