+++ src/test/java/ru/m2mcom/pondnotes/web/rest/ChemicalAnalysisResourceIntTest.java
@@ -64,11 +64,11 @@ public class ChemicalAnalysisResourceIntTest {
     private static final Double DEFAULT_TEMP_VAL = 1D;
     private static final Double UPDATED_TEMP_VAL = 2D;
 
-    private static final Integer DEFAULT_TIMESTAMP = 1;
-    private static final Integer UPDATED_TIMESTAMP = 2;
+    private static final Long DEFAULT_TIMESTAMP = 1L;
+    private static final Long UPDATED_TIMESTAMP = 2L;
 
-    private static final Integer DEFAULT_USER_ID = 1;
-    private static final Integer UPDATED_USER_ID = 2;
+    private static final Long DEFAULT_USER_ID = 1L;
+    private static final Long UPDATED_USER_ID = 2L;
 
     @Autowired
     private ChemicalAnalysisRepository chemicalAnalysisRepository;
@@ -238,8 +238,8 @@ public class ChemicalAnalysisResourceIntTest {
             .andExpect(jsonPath("$.[*].nH4").value(hasItem(DEFAULT_N_H_4.toString())))
             .andExpect(jsonPath("$.[*].ph").value(hasItem(DEFAULT_PH.toString())))
             .andExpect(jsonPath("$.[*].tempVal").value(hasItem(DEFAULT_TEMP_VAL.doubleValue())))
-            .andExpect(jsonPath("$.[*].timestamp").value(hasItem(DEFAULT_TIMESTAMP)))
-            .andExpect(jsonPath("$.[*].userId").value(hasItem(DEFAULT_USER_ID)));
+            .andExpect(jsonPath("$.[*].timestamp").value(hasItem(DEFAULT_TIMESTAMP.intValue())))
+            .andExpect(jsonPath("$.[*].userId").value(hasItem(DEFAULT_USER_ID.intValue())));
     }
 
     @Test
@@ -259,8 +259,8 @@ public class ChemicalAnalysisResourceIntTest {
             .andExpect(jsonPath("$.nH4").value(DEFAULT_N_H_4.toString()))
             .andExpect(jsonPath("$.ph").value(DEFAULT_PH.toString()))
             .andExpect(jsonPath("$.tempVal").value(DEFAULT_TEMP_VAL.doubleValue()))
-            .andExpect(jsonPath("$.timestamp").value(DEFAULT_TIMESTAMP))
-            .andExpect(jsonPath("$.userId").value(DEFAULT_USER_ID));
+            .andExpect(jsonPath("$.timestamp").value(DEFAULT_TIMESTAMP.intValue()))
+            .andExpect(jsonPath("$.userId").value(DEFAULT_USER_ID.intValue()));
     }
 
     @Test
@@ -374,8 +374,8 @@ public class ChemicalAnalysisResourceIntTest {
             .andExpect(jsonPath("$.[*].nH4").value(hasItem(DEFAULT_N_H_4.toString())))
             .andExpect(jsonPath("$.[*].ph").value(hasItem(DEFAULT_PH.toString())))
             .andExpect(jsonPath("$.[*].tempVal").value(hasItem(DEFAULT_TEMP_VAL.doubleValue())))
-            .andExpect(jsonPath("$.[*].timestamp").value(hasItem(DEFAULT_TIMESTAMP)))
-            .andExpect(jsonPath("$.[*].userId").value(hasItem(DEFAULT_USER_ID)));
+            .andExpect(jsonPath("$.[*].timestamp").value(hasItem(DEFAULT_TIMESTAMP.intValue())))
+            .andExpect(jsonPath("$.[*].userId").value(hasItem(DEFAULT_USER_ID.intValue())));
     }
 
     @Test
